Add `PartialOrd` and `Ord` derivations to session value types, optimize comparisons, and enhance page number styles
- Updated `kwh_milli.rs`, `rate.rs`, `grace_period.rs`, `money_yen.rs`, `timeline.rs`, and `bounded.rs` to derive `PartialOrd` and `Ord` for better comparability.
- Simplified `chargeable_energy.rs` logic by leveraging derived ordering for comparisons.
- Improved slide navigation with styled page number display in `styles.css`.

+++ modules/model-b-avdm/src/session/chargeable_energy.rs
@@ -23,7 +23,7 @@ impl ChargeableEnergy {
   /// `billed` が `total` を超えている場合、`SessionValueError::EnergyOutOfRange`
   /// を返します。
   pub fn new(total: KwhMilli, billed: KwhMilli) -> Result<Self, SessionValueError> {
-    if u64::from(billed) > u64::from(total) {
+    if billed > total {
       return Err(SessionValueError::EnergyOutOfRange { provided: u64::from(billed), max: u64::from(total) });
     }
     Ok(Self { total, billed })
